requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
mysql-connector-python==8.4.0
boto3==1.34.113